poc: entity-manager: Introduce adapters for exposed configurations
A goal of entity-manager is to centralise dynamic configuration. System
applications developed in the OpenBMC ecosystem can react to the
published configurations directly, as they participate in the OpenBMC
data model[1] anyway. However, applications developed outside of the
OpenBMC ecosystem are often employed, and also often need configuration.

[1]: https://github.com/openbmc/phosphor-dbus-interfaces

Where configuration of these external applications is possible via
D-Bus, allow the creation of adapters inside entity-manager to directly
drive their interfaces based on exposed entity-manager configurations.

An immediate example of the concept is configuration of MCTP endpoints
via `mctpd`[2]. Implement an adaptor for the MCTPEndpoint schema that
drives application of the configuration-defined EID for the exposed
device.

[2]: https://github.com/CodeConstruct/mctp

Change-Id: I008c80a16ad495a92d0947c41135096e4bd65235
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>